Fix post/term slug syncing when title or slug is updated
Previously, when a title was updated, we would automatically change
the shadow term slug to match that title.

Also previously, when a slug was updated, the change would never
make its way to the shadow term.

With this change, title and slug are both modified as they are
changed with the post.